добавил еще чуть-чуть фич и описание

+++ aiml_bot/main.py
@@ -11,7 +11,7 @@ from google_images_download import google_images_download
 from telegram.ext import Updater, CommandHandler, MessageHandler, Filters
 
 import tts_tools as tts
-from config import TOKEN
+import config
 
 # Enable logging
 logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
@@ -24,21 +24,38 @@ aiml_kernel.learn("std-startup.xml")
 aiml_kernel.respond("load aiml b")
 
 
-def say(text, chat_id, context):
-    output_file = f'audio/{chat_id}_{int(time.time())}'
-    tts.text_to_voice(text, 'ru', output_file)
-    voice = f'{output_file}.ogg'
-    context.bot.send_voice(chat_id=chat_id, voice=open(voice, 'rb'))
-    os.remove(voice)
+def get_function_description():
+    return """С Ботом можно:
+        - /help запросить эту справку
+        - /toggle_voice переключить с голосового режима на текстовый и обратно
+        - Завязать разговор ("привет", "как дела?" и тд)
+        - Переименовать его ("теперь тебя зовут")
+        - Попросить повторить фразу или даже запомнить ее ("повтори ..." и "запомни ...")
+        - Попросить показать что-то ("покажи ...", "покажи мне ...")
+        - Спросить число, дату, день недели, месяц и год ("какое сегодня число", "а год?")
+        - Попросить назвать случайную дату (их список ограничен) и рассказать, что случилось в этот день ("назови случайную дату", "расскажи, что случилось в этот день")
+        - Попросить рассказать что-нибудь ("расскажи что-нибудь")
+    """
+
+
+def say(text, message, context):
+    if config.ANSWER_BY_VOICE:
+        output_file = f'audio/{message.chat.id}_{int(time.time())}'
+        tts.text_to_voice(text, 'ru', output_file)
+        voice = f'{output_file}.ogg'
+        context.bot.send_voice(chat_id=message.chat.id, voice=open(voice, 'rb'))
+        os.remove(voice)
+    else:
+        message.reply_text(text)
 
 
 def start(update, context):
-    update.message.reply_text('Hi!')
+    update.message.reply_text(get_function_description())
+    say(aiml_kernel.respond("привет"), update.message, context)
 
 
 def help(update, context):
-    """Send a message when the command /help is issued."""
-    update.message.reply_text('Help!')
+    update.message.reply_text(get_function_description())
 
 
 def echo(update, context):
@@ -51,8 +68,7 @@ def echo(update, context):
         keywords = search.group()[3:-3]
         subprocess.run(["googleimagesdownload", "--keywords", keywords, "--limit",  "10", "-o", "image/"])
 
-    update.message.reply_text(response)
-    say(response, update.message.chat.id, context)
+    say(response, update.message, context)
     print(f'Sent message "{response}" to {update.message.chat.id}')
     image_path = 'image/'
     for directory in os.listdir(image_path):
@@ -60,6 +76,11 @@ def echo(update, context):
         photos = [os.path.join(dir_path, photo) for photo in os.listdir(dir_path)]
         context.bot.send_photo(chat_id=update.message.chat.id, photo=open(random.choice(photos), 'rb'))
         shutil.rmtree(dir_path)
+
+
+def toggle_voice(update, context):
+    config.ANSWER_BY_VOICE = not config.ANSWER_BY_VOICE
+    say("Готово!", update.message, context)
 
 
 def error(update, context):
@@ -73,7 +94,7 @@ def main():
     # Make sure to set use_context=True to use the new context based callbacks
     # Post version 12 this will no longer be necessary
 
-    updater = Updater(TOKEN, use_context=True)
+    updater = Updater(config.TOKEN, use_context=True)
 
     # Get the dispatcher to register handlers
     dp = updater.dispatcher
@@ -81,6 +102,7 @@ def main():
     # on different commands - answer in Telegram
     dp.add_handler(CommandHandler("start", start))
     dp.add_handler(CommandHandler("help", help))
+    dp.add_handler(CommandHandler("toggle_voice", toggle_voice))
 
     # on noncommand i.e message - echo the message on Telegram
     dp.add_handler(MessageHandler(Filters.text, echo))
